Trying out a Windows-only toolset. Today's latest versions of some spotted and scripted.

+++ tools/create_starter_pack.sh
@@ -1,8 +1,6 @@
 # This script will download external tools and package everything in
 # one zip for workshop participants to download and start creating.
 
-# TODO: Tool binaries for Windows users.
-
 set -e
 
 if [[ $# -lt 1 ]]
@@ -35,16 +33,44 @@ TOOLFILES="tools/pnginator_modified.rb \
   tools/makecompo.bat"
 
 EXTFILES="external/player-small.js \
-          external/closure-compiler.jar"
+          external/closure-compiler.jar \
+	  external/windows-sed.exe \
+	  external/jsexe.exe"
 
 # Download external tools..
-# Closure compiler. Links and names work as of 2019-10-19.
+# Links and names are latest versions as of 2023-03-01.
+
+# Closure compiler is needed for minification on all platforms.
 if [[ ! -e external/closure-compiler.jar ]]
 then
-   curl https://dl.google.com/closure-compiler/compiler-latest.zip > external/compiler-latest.zip
-   unzip -d external/ external/compiler-latest.zip 'closure-*.jar'
-   mv external/closure-*.jar external/closure-compiler.jar
+   curl https://repo1.maven.org/maven2/com/google/javascript/closure-compiler/v20230206/closure-compiler-v20230206.jar > external/closure-compiler.jar
 fi
+
+# Windows users will need sed from somewhere because it is used for some shoveling
+if [[ ! -e external/windows-sed.exe ]]
+then
+    wget -O - https://github.com/mbuilov/sed-windows/raw/master/sed-4.9-x64.exe > external/windows-sed.exe
+fi
+
+# Windows users will get simplest js executable with jsexe
+if [[ ! -e external/jsexe.exe ]]
+then
+    pushd external
+    wget -O - https://files.scene.org/get/resources/tools/windows/adinpsz_-_jsexe.zip > jsexe.zip
+    unzip jsexe.zip jsexe.exe
+    rm jsexe.zip
+    popd
+fi
+
+
+# Tools for building the library itself
+# Shader minifier is needed for minifying the WebGL shaders
+if [[ ! -e external/shader_minifier.exe ]]
+then
+    # Github didn't like default curl so use wget here:
+    wget -O - https://github.com/laurentlb/Shader_Minifier/releases/download/1.3.4/shader_minifier.exe > external/shader_minifier.exe
+fi
+
 
 # Some tools from a forum.. use at your own risk - I leave these in comments..
 #curl https://encode.su/attachment.php?s=639690c9d6d38b7fe4d4200974e8611a&attachmentid=2934&d=1400449686 | tar xvz -O > external/defdb
@@ -62,9 +88,17 @@ cp $LIBFILES "${VERSION}/lib/"
 cp $EXAMPLEFILES "${VERSION}/example/"
 cp $EXTFILES "${VERSION}/external/"
 cp $TOOLFILES "${VERSION}/tools/"
-cp -r documentation "${VERSION}"
-# TODO: this is new
-cp -r winext "${VERSION}"
+
+# Add documentation only if it has been created
+if [[ -e documentation ]]
+then
+  cp -r documentation "${VERSION}"
+else
+    echo "NO DOCUMENTATION FOUND. Skipping."
+fi
+
+# # TODO: this is new .. and not done... I put win stuff in 'external' too
+# cp -r winext "${VERSION}"
 
 
 # Remove earlier version, and re-zip:
